requirements.txt
Flask
flask-cors
psycopg2-binary
kafka-python
gunicorn